Refactor uint FileOffset -> ulong Offset.

+++ src/AsmResolver.PE/Exports/Builder/ExportDirectoryBuffer.cs
@@ -80,7 +80,7 @@ namespace AsmResolver.PE.Exports.Builder
         }
         
         /// <inheritdoc />
-        public override void UpdateOffsets(uint newFileOffset, uint newRva)
+        public override void UpdateOffsets(ulong newFileOffset, uint newRva)
         {
             base.UpdateOffsets(newFileOffset, newRva);
             _contentsBuilder.UpdateOffsets(newFileOffset + ExportDirectoryHeaderSize, newRva + ExportDirectoryHeaderSize);
